fixed camera bug

+++ firmware/lib/XRTL/src/modules/infoLED/InfoLED.h
@@ -101,7 +101,7 @@ public:
 
     void fill(bool fillAll = false);
 
-    void pulse(uint8_t minVal = 40, uint8_t maxVal = 110, uint16_t pulseTime = 1000, uint8_t pulseCount = 2);
+    void pulse(uint8_t minVal = 40, uint8_t maxVal = 110, uint16_t pulseTime = 1000, uint16_t pulseCount = 2);
     void cycle(uint64_t cycleDuration);
 
     void loop();
